requirements.txt
# Core dependencies
pandas
numpy
yfinance
sqlalchemy
psycopg2-binary

# Streaming
kafka-python
pyspark

# Dashboard
streamlit
plotly

# Airflow
apache-airflow
apache-airflow-providers-apache-spark
apache-airflow-providers-postgres

# Monitoring
prometheus_client

# Development
python-dotenv
pytest